Gameplay

+++ src/main/java/com/lineate/testyourlexicon/models/Translation.java
@@ -2,13 +2,12 @@ package com.lineate.testyourlexicon.models;
 
 import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Builder
 @Data
 public class Translation {
 
-  private Integer id;
+  private Long id;
   private String translateFromWord;
   private String translateToWord;
